gerard: a lot of changes.....by Johannes Verelst
The changes:
- The old code had several places where timezones were
hard-coded. DateSupport now uses the local timezone. In several
systems
this results in a wrong offset from GMT (3:30 instead of 2:00)
but this
isn't a problem. It's just a constant, and if it is treated like
a
constant the end-user doesn't notice the difference.

- Documentation
I tried documenting the code in a way that would result in a
readable api
documentation-entry. For all methods (even the small ones) there
now is a
description, and filled-in '@param, @return, @obsolete
and @see' fields.


git-svn-id: https://scm.mmbase.org/mmbase/trunk@1022 91430c5f-d668-0410-9e7d-dd663366dcfd