michiel: the implementation is changed and cleand up heavily. The resulting functionality is nearly completely backwards compatible, but the gain is the following:
- A <mm:context> with an id now is registered in its parent context, this will
allow to treat a context as a variable too (opening the way to passing whole
contextes to another page).

- A context now behaves more exactly like a namespace (if named), or like a
scope (if unnamed). This means that you can refer to objects in the
surrounding contextes now more easily (without importing, or fully
specifying the name). It also allows for better nesting contextes, and referring
to variables in them in a clearer way.  Examples will follow in
mmdemo/examples/taglib.


git-svn-id: https://scm.mmbase.org/mmbase/trunk@2810 91430c5f-d668-0410-9e7d-dd663366dcfd