eduard: Max wasnt remapped and throws an example. Guess what: 	eduard@eduard:~/mmbase/config/default/builders/core$ grep max *.xml 	typerel.xml:     <!-- POS 6 : <field> 'max'  --> 	typerel.xml:          <name xml:lang="en">max</name>
Does this problem also exist in 1.5.1? When this is the case, this would mean that default database doesnt work?

Do we have to research if we could retrieve all the reserved word of the jdbc-connection after the field mapping of the xml, when there are still words, which are reservated, we should also map them.


git-svn-id: https://scm.mmbase.org/mmbase/trunk@4709 91430c5f-d668-0410-9e7d-dd663366dcfd