Some major structural changes on behalf of the cache release strategy project:
- There is an EventManager class, a singleton that takes care of all things concerning events (load/remove
eventbrokers and listeners). This was preveously done in the MMBase class. The cluster manager allso has an
xml file where you can configure what eventbrokers should be loaded at startup (preveously hardcoded).

- The MMBaseChangeInterface has died. It's function has been split up. node changes
are handeled by the ChangeManager, who's only task now is to  wrap them with NodeEvent and RelationEvent instances
and propagates those to the EventManager.

- The ClusterManager is now wwrapped bya module and implements the AllEventListener. It captures remote events of
any type and sends them on to the event manager. It listens to all local events, and sends them into the
cluster. in the module configuration you have to put which specialization of ClusterManager you want to use

- Allso some structuring has been done of the 'etterStrategy (for lack of better name) class, and it has been
added as the default strategy for all QueryResultCache instances.

remarks:
1 the ChangeManager is allmost an empty class. maybe we should dump it?
2 the dtd for the EventManager configuration is not found. I don't know how this should be done(yet)


git-svn-id: https://scm.mmbase.org/mmbase/trunk@15457 91430c5f-d668-0410-9e7d-dd663366dcfd